Update to latest `serenity-next` (#196)
* Companion to serenity-rs/serenity#2523

* Update serenity to include https://github.com/serenity-rs/serenity/pull/2523

* `cargo fmt`

* `cargo update`

Fixes `proc-macro2` compilation failure

* Fix failing tests

* Bump MSRV to 1.72.0

* Update and fix broken things

* Convert u64 to u16

---------

+++ macros/src/modal.rs
@@ -15,8 +15,8 @@ struct StructAttributes {
 struct FieldAttributes {
     name: Option<String>,
     placeholder: Option<String>,
-    min_length: Option<u64>,
-    max_length: Option<u64>,
+    min_length: Option<u16>,
+    max_length: Option<u16>,
     paragraph: Option<()>,
 }
